refactor: move legacy Task files into v1
