+++ src/domains/tasks/features/ListingsPage/components/ListingsPage.tsx
@@ -4,7 +4,7 @@ import axios from "axios";
 import { useSession } from "next-auth/react";
 
 import { Button, Layout, Page } from "@shopify/polaris";
-import { MakeAnOfferModal } from "@Tasks/Offers/features/MakeAnOfferForm";
+import { MakeAnOfferModal } from "@Tasks/MakeAnOfferForm";
 import { ListingCard } from "./ListingCard";
 
 export function ListingsPage({ myTasks }: any) {
