+++ src/domains/tasks/v1/Offers/Dashboard.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { FilterOffers, Offers } from ".";
 import BodyStyles from "@Components/styles/BodyStyles";
-import { TaskDetails } from "@Tasks/TaskDetails";
+import { TaskDetails } from "../TaskDetails";
 
 export function Dashboard() {
   const [selectedTask, setSelectedTask] = useState(null);
