+++ src/domains/tasks/v1/Offers/Offer.tsx
