+++ src/domains/tasks/v1/TaskDetails/Offer.tsx
