+++ src/domains/tasks/v1/TaskDetails/OffersList.tsx
