+++ src/pages/offers.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from "next";
-import { Dashboard } from "@Tasks/Offers";
+import { Dashboard } from "@Tasks/v1/Offers";
 
 const OffersPage: NextPage = () => {
   return <Dashboard />;
